Implement imssage form component
